// src/bell-alert/dto/create-bell-alert.dto.ts
export class CreateBellAlertDto {
  userId: number;

  jobTitleId: number;

  jobTitle: string;

  companyName: string;

  detail: string;

  email: string;

  originLink:string
}

export class BellAlertDto {

  jobId : number;

  userId: number;

  jobTitleId: number;

  jobTitle: string;

  companyName: string;

  detail: string;

  email: string;

  originLink:string
}